requirements.txt
python-dotenv == 0.19.0
fastapi == 0.65.2
uvicorn == 0.15.0
qwhale_client == 0.1.20